+++ R/boot.R
@@ -75,7 +75,7 @@ csboot <- function(model_list, boot_size, block_size, seed = NULL){
 #' @references
 #' Girolimetto, D., Athanasopoulos, G., Di Fonzo, T. and Hyndman, R.J. (2023), Cross-temporal
 #' probabilistic forecast reconciliation: Methodological and practical issues.
-#' \emph{International Journal of Forecasting}, in press. \doi{10.1016/j.ijforecast.2023.10.003}
+#' \emph{International Journal of Forecasting}, 40(3), 1134-1151. \doi{10.1016/j.ijforecast.2023.10.003}
 #'
 #' @family Bootstrap samples
 #' @family Framework: temporal
@@ -135,7 +135,7 @@ teboot <- function(model_list, boot_size, agg_order, block_size = 1, seed = NULL
 #' @references
 #' Girolimetto, D., Athanasopoulos, G., Di Fonzo, T. and Hyndman, R.J. (2023), Cross-temporal
 #' probabilistic forecast reconciliation: Methodological and practical issues.
-#' \emph{International Journal of Forecasting}, in press. \doi{10.1016/j.ijforecast.2023.10.003}
+#' \emph{International Journal of Forecasting}, 40(3), 1134-1151. \doi{10.1016/j.ijforecast.2023.10.003}
 #'
 #' @family Bootstrap samples
 #' @family Framework: cross-temporal