+++ R/cov_cs.R
@@ -22,9 +22,11 @@
 #'      \item "\code{str}" - structural variances.
 #'      \item "\code{wls}" - series variances (uses \code{res}).
 #'      }
-#'     \item Generalized least squares:
+#'     \item Generalized least squares (uses \code{res}):
 #'      \itemize{
-#'      \item "\code{shr}"/"\code{sam}" - shrunk/sample covariance (uses \code{res}).
+#'      \item "\code{shr}" - shrunk covariance (Wickramasuriya et al., 2019).
+#'      \item "\code{oasd}" - oracle shrunk covariance (Ando and Xiao, 2023).
+#'      \item "\code{sam}" - sample covariance.
 #'      }
 #'   }
 #' @param mse If \code{TRUE} (\emph{default}) the residuals used to compute the covariance
@@ -43,7 +45,7 @@
 #' res <- t(matrix(rnorm(n = 30), nrow = 3))
 #'
 #' cov1 <- cscov("ols", n = 3)          # OLS methods
-#' cov2 <- cscov("str", agg_mat = A)   # STR methods
+#' cov2 <- cscov("str", agg_mat = A)    # STR methods
 #' cov3 <- cscov("wls", res = res)      # WLS methods
 #' cov4 <- cscov("shr", res = res)      # SHR methods
 #' cov5 <- cscov("sam", res = res)      # SAM methods
@@ -53,6 +55,10 @@
 #' cscov(comb = "ols2", x = 3) # == cscov("ols", n = 3)
 #'
 #' @references
+#' Ando, S., and Xiao, M. (2023), High-dimensional covariance matrix estimation:
+#' shrinkage toward a diagonal target. \emph{IMF Working Papers}, 2023(257), A001.
+#' \doi{10.5089/9798400260780.001.A001}
+#'
 #' Di Fonzo, T. and Girolimetto, D. (2023), Cross-temporal forecast reconciliation:
 #' Optimal combination method and heuristic alternatives, \emph{International Journal
 #' of Forecasting}, 39, 1, 39-57. \doi{10.1016/j.ijforecast.2021.08.004}
@@ -104,7 +110,7 @@ cscov.str <- function(comb = "str", ..., agg_mat = NULL, strc_mat = NULL){
     strc_mat <- cstools(agg_mat = agg_mat)$strc_mat
   }
 
-  .sparseDiagonal(x = rowSums(strc_mat))
+  .sparseDiagonal(x = rowSums(abs(strc_mat)))
 }
 
 #' @export
@@ -112,8 +118,8 @@ cscov.wls <- function(comb = "wls", ..., res = NULL, mse = TRUE){
   if(is.null(res)){
     cli_abort("Argument {.arg res} is NULL.", call = NULL)
   }
-  res <- na.omit(res)
-  .sparseDiagonal(x = apply(res, 2, function(x) ifelse(mse, sum(x^2)/length(x), var(x))))
+  .sparseDiagonal(x = apply(res, 2, function(x) ifelse(mse, sum(x^2, na.rm = TRUE)/sum(!is.na(x)),
+                                                       var(x, na.rm = TRUE))))
 }
 
 #' @export
@@ -123,6 +129,14 @@ cscov.shr <- function(comb = "shr", ..., res = NULL, mse = TRUE,
     cli_abort("Argument {.arg res} is NULL.", call = NULL)
   }
   shrink_fun(res, mse = mse)
+}
+
+#' @export
+cscov.oasd <- function(comb = "shr", ..., res = NULL, mse = TRUE){
+  if(is.null(res)){
+    cli_abort("Argument {.arg res} is NULL.", call = NULL)
+  }
+  shrink_oasd(res, mse = mse)
 }
 
 #' @export