+++ R/data.R
@@ -16,7 +16,7 @@
 #' @references
 #' Girolimetto, D. and Di Fonzo, T. (2023), Point and probabilistic forecast reconciliation
 #' for general linearly constrained multiple time series,
-#' \emph{Statistical Methods & Applications}, in press. \doi{10.1007/s10260-023-00738-6}.
+#' \emph{Statistical Methods & Applications}, 33, 581-607. \doi{10.1007/s10260-023-00738-6}.
 #'
 #' @source <https://ec.europa.eu/eurostat/web/national-accounts/>
 "itagdp"